prepare talking to anidb

+++ tag-anime-screenshot.py
@@ -5,7 +5,7 @@ import os
 import logging
 import shutil
 import re
-import pyexiv2
+import exiv2
 import configparser
 import argparse
 from sqlalchemy.exc import NoResultFound
@@ -21,12 +21,14 @@ def get_config(config_file=os.path.expanduser("~/.config/weltschmerz/weltschmerz
     parser.add_argument(
         "--database",
         help="database to use",
+        type=str,
         default=config.get("tag-screenshot", "database"),
     )
     parser.add_argument(
         "--log-file",
         dest="log_file",
         help="logfile to use",
+        type=str,
         default=config.get("tag-screenshot", "logfile"),
     )
     parser.add_argument(
@@ -34,16 +36,23 @@ def get_config(config_file=os.path.expanduser("~/.config/weltschmerz/weltschmerz
         "--screenshot-file",
         help="screenshot file to tag",
         default=None,
+        type=str,
         dest="screenshot_file",
     )
     parser.add_argument(
-        "-F", "--file-path", help="file to query for", default=None, dest="file_path"
+        "-F",
+        "--file-path",
+        help="file to query for",
+        type=str,
+        default=None,
+        dest="file_path",
     )
     parser.add_argument(
         "-s",
         "--file-size",
         help="filesize to query for",
         default=None,
+        type=int,
         dest="file_size",
     )
     parser.add_argument("-d", "--duration", help="file duration", default=None)
@@ -65,7 +74,9 @@ def get_config(config_file=os.path.expanduser("~/.config/weltschmerz/weltschmerz
         default=config.get("tag-screenshot", "basedir"),
         dest="target_basedir",
     )
-    parser.add_argument("--screenshot-type", help="screenshot type", default=1)
+    parser.add_argument(
+        "--screenshot-type", help="screenshot type", type=int, default=1
+    )
 
     args = parser.parse_args()
     logging.basicConfig(
@@ -77,23 +88,24 @@ def get_config(config_file=os.path.expanduser("~/.config/weltschmerz/weltschmerz
 
 
 def tag_screenshot(
-    screenshot,
-    filename,
-    filesize,
-    duration,
-    duration_raw,
-    timepos,
-    timepos_raw,
-    screenshot_type,
-    hash_crc,
-    hash_md5,
-    hash_sha1,
-    hash_ed2k,
+    screenshot: str,
+    filename: str,
+    filesize: int,
+    duration: str,
+    duration_raw: str,
+    timepos: str,
+    timepos_raw: str,
+    screenshot_type: int,
+    hash_crc: str,
+    hash_md5: str,
+    hash_sha1: str,
+    hash_ed2k: str,
 ):
     logging.info(f"tagging {config.screenshot_file}...")
-    pyexiv2.xmp.register_namespace("http://ns.eroforce.one/anime/1.0/", "anime")
-    metadata = pyexiv2.ImageMetadata(screenshot)
-    metadata.read()
+    exiv2.XmpProperties.registerNs("http://ns.eroforce.one/anime/1.0/", "anime")
+    screenshot_image = exiv2.ImageFactory.open(screenshot)
+    screenshot_image.readMetadata()
+    metadata = screenshot_image.xmpData()
     metadata["Xmp.anime.AnimeFileName"] = filename
     metadata["Xmp.anime.AnimeFileSize"] = filesize
     metadata["Xmp.anime.AnimeFileDuration"] = duration
@@ -105,30 +117,35 @@ def tag_screenshot(
     metadata["Xmp.anime.AnimeFileSha1"] = hash_sha1
     metadata["Xmp.anime.AnimeFileEd2k"] = hash_ed2k
     metadata["Xmp.anime.ScreenshotType"] = screenshot_type
-    metadata.write()
+    screenshot_image.setXmpData(metadata)
+    screenshot_image.writeMetadata()
     logging.info(f"tagged {config.screenshot_file}")
 
 
 def tag_screenshot_min(
-    screenshot,
-    filename,
-    filesize,
-    duration,
-    duration_raw,
-    timepos,
-    timepos_raw,
-    screenshot_type,
+    screenshot: str,
+    filename: str,
+    filesize: int,
+    duration: str,
+    duration_raw: str,
+    timepos: str,
+    timepos_raw: str,
+    screenshot_type: int,
 ):
     logging.info(f"tagging (min) {config.screenshot_file}...")
-    pyexiv2.xmp.register_namespace("http://ns.eroforce.one/anime/1.0/", "anime")
-    metadata = pyexiv2.ImageMetadata(screenshot)
-    metadata.read()
+    exiv2.XmpProperties.registerNs("http://ns.eroforce.one/anime/1.0/", "anime")
+    screenshot_image = exiv2.ImageFactory.open(screenshot)
+    screenshot_image.readMetadata()
+    metadata = screenshot_image.xmpData()
+    metadata["Xmp.anime.AnimeFileName"] = filename
+    metadata["Xmp.anime.AnimeFileSize"] = filesize
     metadata["Xmp.anime.AnimeFileDuration"] = duration
     metadata["Xmp.anime.AnimeFileDurationRaw"] = duration_raw
     metadata["Xmp.anime.AnimeFileTime"] = timepos
     metadata["Xmp.anime.AnimeFileTimeRaw"] = timepos_raw
     metadata["Xmp.anime.ScreenshotType"] = screenshot_type
-    metadata.write()
+    screenshot_image.setXmpData(metadata)
+    screenshot_image.writeMetadata()
     logging.info(f"tagged (min) {config.screenshot_file}")
 
 
@@ -150,7 +167,7 @@ if __name__ == "__main__":
             .one()
         )
         logging.info(
-            f"found file '{filename}' in DB, adding metadata for {config.timepos}..."
+            f"found file '{filename}' in DB, adding metadata for {config.timepos} ..."
         )
         tag_screenshot(
             screenshot=config.screenshot_file,
@@ -180,26 +197,27 @@ if __name__ == "__main__":
             time_position_raw=config.timepos_raw,
             title_type=config.screenshot_type,
         )
-        target_folder = os.path.join(
-            config.target_basedir,
-            re.search("(by-id(/[0-9]{2}){3})", real_path).group(1),
-        )
-        target_file = os.path.join(
-            target_folder, os.path.basename(config.screenshot_file)
-        )
-        if os.path.isfile(target_file):
-            logging.warning(
-                f"screenshot '{target_file}' already present, deleting '{config.screenshot_file}'..."
+        if re.match("(by-id(/[0-9]{2}){3})", real_path):
+            target_folder = os.path.join(
+                config.target_basedir,
+                re.search("(by-id(/[0-9]{2}){3})", real_path).group(1),
             )
-            os.unlink(config.screenshot_file)
-        else:
-            logging.info(f"found anime id, moving file to '{target_folder}'...")
-            os.makedirs(os.path.abspath(target_folder), exist_ok=True)
-            shutil.move(config.screenshot_file, target_file)
+            target_file = os.path.join(
+                target_folder, os.path.basename(config.screenshot_file)
+            )
+            if os.path.isfile(target_file):
+                logging.warning(
+                    f"screenshot '{target_file}' already present, deleting '{config.screenshot_file}' ..."
+                )
+                os.unlink(config.screenshot_file)
+            else:
+                logging.info(f"found anime id, moving file to '{target_folder}'...")
+                os.makedirs(os.path.abspath(target_folder), exist_ok=True)
+                shutil.move(config.screenshot_file, target_file)
+            tss.filename = target_file
         logging.info(
-            f"adding title screenshot to DB for '{target_file}' at {config.timepos}..."
+            f"adding title screenshot to DB for '{tss.filename}' at {config.timepos} ..."
         )
-        tss.filename = target_file
         try:
             anidb_file = (
                 dbs.session.query(anime.File)
@@ -224,4 +242,4 @@ if __name__ == "__main__":
             timepos_raw=config.timepos_raw,
             screenshot_type=config.screenshot_type,
         )
-        logging.warning(f"'{filename}' not found in DB, only tagged minimal info...")
+        logging.warning(f"'{filename}' not found in DB, only tagged minimal info ...")
